requirements.txt
python-pptx>=0.6.23
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.1
moviepy>=1.0.3,<2.0.0
markdown-it-py>=3.0.0
azure-identity>=1.17.0
